Split song_state.h god-file: extract EnergyState, GameOverState, DeathCause (refs #1194)
Per the components/ audit, song_state.h bundled four unrelated singletons
across different ownership lifecycles. Splits along the natural seams:

| Type               | New home          | Concern                       |
|--------------------|-------------------|-------------------------------|
| SongState          | song_state.h      | timing/playback state         |
| SongResults        | song_state.h      | song-scope result tallies     |
| EnergyState        | energy_bar.h      | gameplay-resource singleton   |
| GameOverState      | game_state.h      | terminal-phase data           |
| DeathCause         | game_state.h      | terminal-phase enum           |

energy_bar.h now co-locates the gameplay-resource singleton with the HUD
visual contract (EnergyBarLayout / EnergyBarVisual) for the same concept.
game_state.h gains the terminal-cause that pairs with GamePhase::GameOver.

rhythm.h (back-compat re-export shim) updated to also include energy_bar.h
so all four runtime singletons remain reachable through the legacy header
without forcing every consumer to discover the new locations on this pass.

Direct includes patched only where consumers include song_state.h directly
and used moved types: game_over_screen_controller.h (DeathCause →
game_state.h), song_complete_screen_controller.cpp (EnergyState →
energy_bar.h), play_session.cpp (EnergyState → energy_bar.h).

Validation: 902 Catch2 tests / 2957 assertions pass; zero warnings under
both unity and non-unity Clang -Wall -Wextra -Werror builds.
