+++ app/components/rhythm.h
@@ -2,11 +2,13 @@
 
 // rhythm.h — Per-entity rhythm components and timing enums.
 // Loaded-once data (BeatMap) lives in beat_map.h.
-// Runtime singletons (SongState, EnergyState, SongResults) live in song_state.h.
-// This header re-exports both for backward compatibility.
+// Runtime singletons: SongState / SongResults live in song_state.h;
+// EnergyState lives in energy_bar.h. This header re-exports all three for
+// backward compatibility.
 
 #include "beat_map.h"
 #include "song_state.h"
+#include "energy_bar.h"
 #include "player.h"
 #include "obstacle.h"
 #include <cstdint>
